Implement MonitorImplX11
