--- src/SFML/Window/MonitorImpl.cpp
@@ -1,98 +1,0 @@
-////////////////////////////////////////////////////////////
-//
-// SFML - Simple and Fast Multimedia Library
-// Copyright (C) 2007-2024 Laurent Gomila (laurent@sfml-dev.org)
-//
-// This software is provided 'as-is', without any express or implied warranty.
-// In no event will the authors be held liable for any damages arising from the use of this software.
-//
-// Permission is granted to anyone to use this software for any purpose,
-// including commercial applications, and to alter it and redistribute it freely,
-// subject to the following restrictions:
-//
-// 1. The origin of this software must not be misrepresented;
-//    you must not claim that you wrote the original software.
-//    If you use this software in a product, an acknowledgment
-//    in the product documentation would be appreciated but is not required.
-//
-// 2. Altered source versions must be plainly marked as such,
-//    and must not be misrepresented as being the original software.
-//
-// 3. This notice may not be removed or altered from any source distribution.
-//
-////////////////////////////////////////////////////////////
-
-////////////////////////////////////////////////////////////
-// Headers
-////////////////////////////////////////////////////////////
-#include <SFML/Window/MonitorImpl.hpp>
-#include <SFML/Config.hpp>
-
-#include <memory>
-
-
-#if defined(SFML_SYSTEM_WINDOWS)
-
-#include <SFML/Window/Win32/MonitorImplWin32.hpp>
-using MonitorImplType = sf::priv::MonitorImplWin32;
-
-#elif defined(SFML_SYSTEM_LINUX) || defined(SFML_SYSTEM_FREEBSD) || defined(SFML_SYSTEM_OPENBSD) || \
-    defined(SFML_SYSTEM_NETBSD)
-
-#if defined(SFML_USE_DRM)
-
-#include <SFML/Window/DRM/MonitorImplDRM.hpp>
-using MonitorImplType = sf::priv::MonitorImplDRM;
-
-#else
-
-#include <SFML/Window/Unix/MonitorImplX11.hpp>
-using MonitorImplType = sf::priv::MonitorImplX11;
-
-#endif
-
-#elif defined(SFML_SYSTEM_MACOS)
-
-#include <SFML/Window/macOS/MonitorImplCocoa.hpp>
-using MonitorImplType = sf::priv::MonitorImplCocoa;
-
-#elif defined(SFML_SYSTEM_IOS)
-
-#include <SFML/Window/iOS/MonitorImplUIKit.hpp>
-using MonitorImplType = sf::priv::MonitorImplUIKit;
-
-#elif defined(SFML_SYSTEM_ANDROID)
-
-#include <SFML/Window/Android/MonitorImplAndroid.hpp>
-using MonitorImplType = sf::priv::MonitorImplAndroid;
-
-#endif
-
-
-namespace sf::priv
-{
-
-////////////////////////////////////////////////////////////
-std::unique_ptr<MonitorImpl>&& MonitorImpl::getPrimaryMonitor()
-{
-    return MonitorImplType::getPrimaryMonitor();
-}
-
-
-////////////////////////////////////////////////////////////
-std::vector<VideoMode> MonitorImpl::getFullscreenModes() {
-	return static_cast<MonitorImplType*>(this)->getFullscreenModes();
-}
-
-
-////////////////////////////////////////////////////////////
-/// \brief Get the current desktop video mode
-///
-/// \return Current desktop video mode
-///
-////////////////////////////////////////////////////////////
-VideoMode MonitorImpl::getDesktopMode() {
-	return static_cast<MonitorImplType*>(this)->getFullscreenModes();
-}
-
-} // namespace sf::priv
