+++ src/SFML/Window/Win32/MonitorImplWin32.cpp
